TAO-8656 -  move TestDefinitionSerializerServiceTest to unit tests
